03_agents/00_agency_swarm_framework/requirements.txt
# Using older version since OpenAI has updated assistants to V2 that is not compatible with Azure OpenAI
# https://github.com/VRSEN/agency-swarm/issues/123
# agency-swarm==0.1.7

# Latest version(0.2.6) has parallel_tool_calls enabled which is not supported by Azure OpenAI.
# Azure OpenAI is not keeping up with the OpenAI API changes which is making it difficult to use this framework's
# latest version.
agency-swarm==0.2.5
tavily-python
duckduckgo-search
python-dotenv
# For windows based systems
pyreadline3